Define method combinators in a uniform way.
All method combinators are now defined as a `newtype` in terms
of a single `Method` combinator. This is mostly backwards-compatible
(except that the choice of the name `Method` clashes for some
interpretations). However, it provides the option to interpretations
to define the `Method` cases in a uniform way by giving a single
instance for `Method` and then using newtype-deriving to derive all
the specific cases.