Adapting the server interpretation.
A wart is that generalized newtype deriving doesn't work for classes
with associated type synonyms (cf. GHC bug 4083, which mentions it
might work for type synonyms, but is conservatively disabled right
now).

So we have to manually define trivial instances for all the newtype
wrappers. It still seems an improvement though.